Add files via upload
This is my journey of learning the python coding language this first one is very simple. It asks for a name and age and adds 100 years to the age and prints the new age out. Also as a bonus i ask for a number and repeat the string that many times as the number that was input.